refactor of newsitem
